+++ packages/button/src/PrimaryButton.tsx
@@ -1,3 +1,5 @@
+import React from 'react';
+
 import { BoxAs, BoxProps, BoxStyleProps } from '@realsystem/box';
 import styled from '@realsystem/styling';
 import { merge } from '@realsystem/utils';