+++ packages/button/stories/Button.stories.tsx
@@ -1,3 +1,4 @@
+import React from 'react';
 import { Meta } from '@storybook/react/types-6-0';
 
 import { Button } from '@realsystem/core/button';