+++ tools/dev.js
@@ -1,5 +1,5 @@
 const esbuild = require('esbuild');
-const { logger, makeEsbuildConfig, getFieldsFromPkgJson } = require('./utils');
+const { logger, makeEsbuildConfig } = require('./utils');
 
 /**
  * @function dev bundle package for cjs and esm output
@@ -7,29 +7,24 @@ const { logger, makeEsbuildConfig, getFieldsFromPkgJson } = require('./utils');
  * @todo update onrebuild in each package
  */
 function dev(packageJson) {
-  const { outfile, ...pkgJson } = getFieldsFromPkgJson(packageJson);
-  const cjsConfig = makeEsbuildConfig(pkgJson, {
-    format: 'cjs',
-    outfile: outfile.cjs,
-  });
-  const esmConfig = makeEsbuildConfig(pkgJson, {
+  const cjsConfig = makeEsbuildConfig(packageJson);
+  const esmConfig = makeEsbuildConfig(packageJson, {
     format: 'esm',
-    outfile: outfile.esm,
   });
 
-  logger.magenta(`Bundling ${pkgJson.name}\n`);
+  logger.magenta(`Bundling ${packageJson.name}\n`);
 
   // bundle commonjs
   esbuild
     .build(cjsConfig)
-    .then(async () => logger.success(`Bundled CJS for "${pkgJson.name}"`))
+    .then(async () => logger.success(`Bundled CJS for "${packageJson.name}"`))
     .catch(() => process.exit(1));
 
   // bundle esmodule
   esbuild
     .build(esmConfig)
     .then(async () => {
-      logger.success(`Bundled ESM for "${pkgJson.name}"\n`);
+      logger.success(`Bundled ESM for "${packageJson.name}"\n`);
       logger.blue('-------------------------------------------------- \n');
     })
     .catch(() => process.exit(1));